fix(compiler): split argument resolvers
