+++ packages/compiler/packages/tokenizer/src/syntax/parseArgument/classifyIntermodularReference.ts
@@ -1,0 +1,132 @@
+import type { ArgumentIdentifier } from '@8f4e/compiler-spec';
+import { ArgumentType } from '@8f4e/compiler-spec';
+
+/** Classifies intermodular module-base and memory-reference identifier forms. */
+export function classifyIntermodularReference(value: string): ArgumentIdentifier | null {
+	if (value.startsWith('&')) {
+		return classifyStartAddressIntermodularReference(value);
+	}
+
+	if (value.endsWith('&')) {
+		return classifyEndAddressIntermodularReference(value);
+	}
+
+	return null;
+}
+
+function classifyStartAddressIntermodularReference(value: string): ArgumentIdentifier | null {
+	const colonIndex = value.indexOf(':', 1);
+	if (colonIndex === -1 || value.indexOf(':', colonIndex + 1) !== -1) {
+		return null;
+	}
+
+	const targetModuleId = value.slice(1, colonIndex);
+	if (!isValidModuleReferenceName(targetModuleId)) {
+		return null;
+	}
+
+	const targetMemoryId = value.slice(colonIndex + 1);
+	if (targetMemoryId.length === 0) {
+		return {
+			type: ArgumentType.IDENTIFIER,
+			value,
+			referenceKind: 'intermodular-module-reference',
+			scope: 'intermodule',
+			targetModuleId,
+			isEndAddress: false,
+		};
+	}
+
+	if (isDecimalDigits(targetMemoryId)) {
+		return {
+			type: ArgumentType.IDENTIFIER,
+			value,
+			referenceKind: 'intermodular-module-nth-reference',
+			scope: 'intermodule',
+			targetModuleId,
+			targetMemoryIndex: Number.parseInt(targetMemoryId, 10),
+		};
+	}
+
+	if (!isValidModuleReferenceName(targetMemoryId)) {
+		return null;
+	}
+
+	return {
+		type: ArgumentType.IDENTIFIER,
+		value,
+		referenceKind: 'intermodular-reference',
+		scope: 'intermodule',
+		targetModuleId,
+		targetMemoryId,
+		isEndAddress: false,
+	};
+}
+
+function classifyEndAddressIntermodularReference(value: string): ArgumentIdentifier | null {
+	const colonIndex = value.indexOf(':');
+	if (colonIndex === -1 || value.indexOf(':', colonIndex + 1) !== -1) {
+		return null;
+	}
+
+	const targetModuleId = value.slice(0, colonIndex);
+	if (!isValidModuleReferenceName(targetModuleId)) {
+		return null;
+	}
+
+	const targetMemoryId = value.slice(colonIndex + 1, -1);
+	if (targetMemoryId.length === 0) {
+		return {
+			type: ArgumentType.IDENTIFIER,
+			value,
+			referenceKind: 'intermodular-module-reference',
+			scope: 'intermodule',
+			targetModuleId,
+			isEndAddress: true,
+		};
+	}
+
+	if (!isValidModuleReferenceName(targetMemoryId)) {
+		return null;
+	}
+
+	return {
+		type: ArgumentType.IDENTIFIER,
+		value,
+		referenceKind: 'intermodular-reference',
+		scope: 'intermodule',
+		targetModuleId,
+		targetMemoryId,
+		isEndAddress: true,
+	};
+}
+
+function isValidModuleReferenceName(value: string): boolean {
+	if (value.length === 0) {
+		return false;
+	}
+
+	for (let index = 0; index < value.length; index++) {
+		const char = value[index];
+		if (char === '&' || char === ':' || char === '.' || char.trim() === '') {
+			return false;
+		}
+	}
+
+	return true;
+}
+
+function isDecimalDigits(value: string): boolean {
+	if (value.length === 0) {
+		return false;
+	}
+
+	for (let index = 0; index < value.length; index++) {
+		const charCode = value.charCodeAt(index);
+		if (charCode < 48 || charCode > 57) {
+			return false;
+		}
+	}
+
+	return true;
+}
